requirements.txt
pandas==2.2.2
numpy==1.26.4
scipy==1.13.1
scikit-learn==1.5.1
PyYAML==6.0.2
matplotlib==3.9.0
tqdm==4.66.5
